requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
gunicorn==22.0.0
mysql-connector-python==9.0.0
